+++ src/main/java/com/restaurants/demo/dto/response/OrderResponse.java
@@ -19,6 +19,6 @@ public class OrderResponse {
     private Integer tableNumber;
     private OrderStatus status;
     private List<OrderItemResponse> items;
-    private BigDecimal totalAmount;
+    private Long totalAmount;
     private LocalDateTime createdAt;
 }
